Function for calculating theoretical joint probabilities at certain distance

Using assumption that all non-background footprints have the same emission probabilities we calculate theoretical probabilities P(0,0|S), P(0,1|S), P(1,0|S), P(1,1|S) for all distances S up to max_spacing. P(0,0|S) etc, are probabilities to find 0 and 0 separated by distance S.

calculate_theor_joint_probs(
  ftp_cover_priors,
  bg_protect_prob,
  footprint_protect_prob,
  max_spacing
)

Arguments

ftp_cover_priors
coverage probabilities for each footprint. IMPORTANT! This must be a vector where ftp_cover_priors[1] is coverage probability of background and we assume that indexes in the vector correspond to the footprint lengths.

bg_protect_prob
emission probability for background for 1.

footprint_protect_prob
emission probability for footprints for 1.

max_spacing
maximum distance

Value

data.frame containing columns: S - distance where S = 1 corresponds to the same position and probabilities are marginal probabilities to observe 0 and 1; P00, P01, P10, P11 - calculated theoretical P(0,0|S), P(0,1|S), P(1,0|S), P(1,1|S).