Title

Title

fit_background_mixmodel_from_aggreg_table(
  protect_stat_table,
  method = c("optimizing", "sampling", "vb"),
  alpha_distrbetaprior_mixcoeff = 1,
  beta_distrbetaprior_mixcoeff = 0.001,
  alpha_distrbetaprior_bg_protect = 0.1,
  beta_distrbetaprior_bg_protect = 1,
  nchains = 4,
  ncpu = nchains,
  ...
)

Arguments

protect_stat_table
matrix with columns TotalGCH - total number of GCHs in a fragment, ProtectedGCH - number of protected GCHs in a fragments, Nfrags - number of fragments with combination of TotalGCH and ProtectedGCH.

method
method for parameter inference that calls corresponding rstan function. optimizing returns point estimate corresponding to maximum in posterior denstiy. sampling returns Hamiltonian Monte Carlo draws from posterior distribution. vb return draws from ADVI approximation of posterior distribution (not tested yet).

alpha_distrbetaprior_mixcoeff
alpha (shape1 in rbeta) parameter for beta prior distribution for mixing coefficient.

beta_distrbetaprior_mixcoeff
beta (shape2 in rbeta) parameter for beta prior distribution for mixing coefficient.

alpha_distrbetaprior_bg_protect
alpha (shape1 in rbeta) parameter for beta prior distribution for bg_protect_prob.

beta_distrbetaprior_bg_protect
beta (shape2 in rbeta) parameter for beta prior distribution for bg_protect_prob.

nchains
number of markov chains for rstan::sampling. Currently has no effect

ncpu
number of cpu to use for rstan::sampling. Currently has no effect

...
parameters for rstan::optimizing

Value

a list containing following slots: rstan_output_object rstan objected returned by function optimizing. background_model_params list containing parameters for function which performs footprint inference, i.e. nomeR::infer_footprints_stan_sampling. The list background_model_params contains 3 values: bg_protect_prob_fixed inferred value for background probability to emit protected base, i.e. bg_protect_prob. bg_protect_prior_alpha alpha (shape1 in rbeta) parameter for beta distribution which will be used as a prior distribution for bg_protect_prob in function for footprint inference. bg_protect_prior_beta beta (shape2 in rbeta) parameter for beta distribution which will be used as a prior distribution for bg_protect_prob in function for footprint inference.