Fetch statistics of duplicated fragments

Fetch statistics of duplicated fragments

get_dupl_stats_from_bams(
  bamfiles,
  samplenames,
  regions,
  genome,
  collapseBySample = TRUE,
  mapqMin = 0,
  mapqMax = 255,
  max_read_size = 1000L,
  ncores = 1L
)

Arguments

bamfiles
character paths to bam files

samplenames
character names of samples. If collapseBySample is TRUE counts are aggregated across bam files with same sample name.

regions
GRanges object for regions of interest.

genome
character path to fasta file which was used as reference for alignment.

collapseBySample
logical indicating whether to collapse counts for bam files with same samplenames. If FALSE prefix s followed by index is added to samplenames.

mapqMin
integer setting minimum MAPQ of reads to be included into analysis.

mapqMax
integer setting maximum MAPQ of reads to be included into analysis.

max_read_size
maximum read length.

ncores
number of cores to use.

Value

tibble where each row corresponds to sample - region combination. Columns of the returned tibble represent:

SampleName

Name of a sample as defined by samplenames.

seqnames, start, end

seqnames (or chromosomes) and coordinates of corresponding regions.

names

the same as names(regions) if it is not NULL or index of a corresponding region in regions.

nFragsFetched

number of fetched fragments before filtering.

nFragsNonUnique

number of non-unique fragments as defined by coordinates and methylation profiles.

duplStatsMatrix

matrix with 2 columns. First column represents (n_uniq_aln) number of unique alignments after deduplication and second column (n_times_duplicated) represents number of times each of the alignments was duplicated. For example, row with numbers 100 at the first column and 1 at the second column should be interpreted as following: 100 alignments in the data has exactly 1 unique fragment. In other words, 100 fragments are unique. Another example, row with numbers 200 at the first column and 5 at the second column means that before deduplication 5 alignments were supported by 200 duplicated fragments each.

duplFragNamesList

list of vectors containing names of duplicated fragments. Names of list elements are encodings of unique alignments and can be ignored.