Extract and/or plot estimates from footprint inference and identify potential footprints from spectrum

This utility function extract estimates from infer_stanfit and uses function suggest_footprints to detect peaks in spectrum and return lengths of potential footprints.

get_ftp_inference_summary(
  infer_stanfit,
  ftp_abundance_name = c("ftp_cover_probs", "ftp_start_probs"),
  plot = F,
  plot_posterior_range = c("2.5%", "97.5%"),
  max_peak_width = 30,
  spline_spar = 0.6,
  max_abund_log2drop = 1.5,
  ...
)

Arguments

infer_stanfit
stanfit object returned by infer_footprints_stan_sampling, infer_footprints_stan_vb or a list returned by infer_footprints_stan_optimizing

ftp_abundance_name
character which defines which footprint abundance value to return/display. Currently available "ftp_cover_probs" or "ftp_start_probs" which corresponds to footprint coverage probabilities of footprint start probabilities.

plot
logical plot footprint abundance spectrum

plot_posterior_range
vector of character of length 2 which specifies which posterior range to plot. summary,stanfit-method provides "2.5%", "25%" and "75%", "97.5%" credible intervals as well as SD and SE. Displaying SD and SE is not implemented at the moment.

max_peak_width
parameter for function suggest_footprints, namely maximum width of detected peaks in spectrum. Please see suggest_footprints.

spline_spar
parameter for function suggest_footprints controlling smoothness of spline ((0,1], the higher the smoother). Please see suggest_footprints smooth.spline. It is recommended to test different values for spline_spar, for example 0.1, 0.3, 0.5, 0.75 to check whether suggested footprints look as expected.

max_abund_log2drop
parameter for function suggest_footprints,namely maximum decrease in abundance relative to value at local maxima until which peaks are extended. Please see suggest_footprints

...
parameters for suggest_footprints which are transmitted to function smooth.spline.

Value

list containing elements:

"ESTIMATES" - list which contains a data.frame for footprint abundances estimates, "footprint_protect_prob_estimate" and bg_protect_prob_estimate which are estimates for footprint_protect_prob and bg_protect_prob. "bg_protect_prob_estimate" is NA if inference for this parameter is not available.

"FTP_SUGGEST" - matrix with coordinates for suggested footprints as returned by suggest_footprints.

"PLOT" - a ggplot object with plot for footprint spectrum.

Examples

if (FALSE) {
 
## simple data with two ftp of 5 and 10 bps. 
## The table below is a count table of observed occurrences of 
## 00, 01, 10 and 11 at spacings from 1 until 15.
 
ftp_5_10_data <- data.frame("S" = 1:15,
                            "N00" = c(1626964,1508381,
                                      1420066,1336897,
                                      1258679,1185045,
                                      1136784,1090029,
                                      1045066,1001670,
                                      959927,983020,
                                      1000410,1012326,
                                      1019633),
                            "N01" = c(0,113856,
                                      197657,276539,
                                      350664,420399,
                                      464873,507988,
                                      549466,589487,
                                      627997,601629,
                                      580965,565776,
                                      555217),
                            "N10" = c(0, 113921,
                                      197854,276862,
                                      351147,421042,
                                      465716,509005,
                                      550645,590837,
                                      629533,603328,
                                      582814,567737,
                                      557220),
                            "N11" = c(873036,758842,
                                      674423,594702,
                                      519510,448514,
                                      402627,357978,
                                      314823,273006,
                                      232543,257023,
                                      275811,289161,
                                      297930))

## variational inference for footprints in the data
hmc_output <- infer_footprints_stan_sampling(joint_freq_table = ftp_5_10_data,
                                      footprint_prior_diralphas = c(10,rep(1,14)))
                                      
## get estimates and plot footprint spectrum
inference_summary_list <- get_inference_summary(hmc_output,plot=T)

}