Create parameters for predict_footprints using footprints summary returned by get_ftp_inference_summary

This utility function creates a list of parameters required by predict_footprints to predict footprint positions in data.

get_ftp_models_for_prediction(infer_summary)

Arguments

infer_summary
list returned by get_ftp_inference_summary which must contain "ESTIMATES" and "FTP_SUGGEST". If case you are not satisfied with footprints suggested by get_ftp_inference_summary substitute slot "FTP_SUGGEST" with your own n x 2 matrix.

Value

list containing

"FTP_MODELS" - footprint models corresponding to parameter footprint_models in predict_footprints.

"bgprotectprob" - numeric background protection probability corresponding to parameter bgprotectprob in predict_footprints. "bgcoverprior" - numeric background protection probability corresponding to parameter bgcoverprior in predict_footprints.