Fetch protection aggregated statistics from NOMe-seq BAM files

Function retrieves statistics of number of protected GCHs out of total number of GCHs in fragments in a set of BAM files generated for NOMe-seq data and set of regions of interest. This statistics is mainly needed for analysis of NOMe-seq data generated for naked DNA controls, such as lambda DNA, and used for estimating noise in background and create background model for later use during footprint inference or footprint prediction.

get_protect_stats_table_from_bams(
  bamfiles,
  samplenames,
  regions,
  genome,
  collapseBySample = TRUE,
  remove_nonunique = TRUE,
  clip_until_nbg = 1L,
  noclip_protect_frac_above = 0.9,
  max_bisC_meth = 0.1,
  min_bisC_size = 10,
  mapqMin = 0,
  mapqMax = 255,
  max_read_size = 1000L,
  ncores = 1L
)

Arguments

bamfiles
character paths to bam files

samplenames
character names of samples. If collapseBySample is TRUE counts are aggregated across bam files with same sample name.

regions
GRanges object for regions of interest.

genome
character path to fasta file which was used as reference for alignment.

collapseBySample
logical indicating whether to collapse counts for bam files with same samplenames. If FALSE prefix s followed by index is added to samplenames.

remove_nonunique
logical if TRUE only unique fragments are analyzed. Uniqueness is defined by fragments start, end and methylation states of all cytosines.

clip_until_nbg
integer controlling clipping partial footprints at both ends of fragments. Namely, protected states are erased from each end until clip_until_nbg unprotected states are met. Set clip_until_nbg=0L to skip clipping preprocessing.

noclip_protect_frac_above
numeric controlling fragment clipping, i.e. fragments which have fraction of protected states higher than max_protect_frac escape clipping. This parameter allows to avoid removal of fully protected fragments during clipping process.

max_bisC_meth
numeric controlling removal of fragments with failed bisulfite conversion, i.e. fragments with fraction of methylated non-GCH, and non-WCG cytosines higher then max_bisC_meth are removed from the analysis to get rid of fragments for which bisulfite conversion failed.

min_bisC_size
integer setting minimum number non-GCH, and non-WCG cytosines for filtering according to bisulfite conversion efficiency, i.e. only fragments with number of non-GCH, and non-WCG cytosines higher then min_bisC_size are subject to filtering based on bisulfite conversion efficiency.

mapqMin
integer setting minimum MAPQ of reads to be included into analysis.

mapqMax
integer setting maximum MAPQ of reads to be included into analysis.

max_read_size
maximum read length.

ncores
number of cores to use.

Value

tibble where each row corresponds to sample - region combination. Columns of the returned tibble represent:

SampleName

Name of a sample as defined by samplenames

.
seqnames, start, end

seqnames (or chromosomes) and coordinates of corresponding regions.

names

the same as names(regions) if it is not NULL or index of a corresponding region in regions.

nFragsFetched

number of fetched fragments before filtering.

nFragsNonUnique

number of non-unique fragments as defined by coordinates and methylation profiles.

nFragsBisFailed

number of fragments for which bisulfite conversion may have failed, namely non-GCH and non-WCG methylation level is higher than max_bisC_meth.

nFragsAnalyzed

number of fragments analysed for co-occurrence table after filtering non-unique and bisulfite conversion failed fragments.

ClippedUntilNbg

parameter clip_until_nbg used for clipping partial footprints from both ends.

NotClippedProtectAbove

parameter noclip_protect_frac_above used for clipping partial footprints from both ends.

nFragsClipped

number of fragments clipped from both ends to get rid of paetial footprints.

ProtectStats

matrix with columns reflecting combinations of total number and number of protected GCHs:

TotalGCH

total number of GCHs.

ProtectedGCH

number of of protected GCHs.

Nfrags

number of fragments met with combination of TotalGCH and ProtectedGCH.