nomeR: Probabilistic model for NOME-Seq data analysis

This packages uses a probabilistic model for analysis of NOMe-Seq data, or any other kind of data that gives single molecule occupancy read out.

Details

The model allows to perform two types of analysis.

1. Inference of footprint abundances in NOMe-seq data

This analysis can be done when the main question is what footprints are present in the data. By counting occurences of observing combinations of protected/unprotected positions at all possible distances and using a bayesian model we make avalilable two algorithms for footprint inference.

Sampling from posterior distribution using Monte Carlo methods from `rstan` package (infer_footprints_stan_sampling);

Approximation of posterior distribution using variational Bayes from `rstan` package (infer_footprints_stan_vb).

Finding maximum a posteriori point estimate using optimization algorithms provided by 'rstan' (infer_footprints_stan_optimizing).

2. Prediction of footprint positions in NOMe-seq data

When footprint abundances are known or inferred from the analysis in previous section one can predict positions of footprints in the NOMe-seq data (predict_footprints). We calculate posterior probabilities for each footprint to start or to cover each position in each fragment.