Function for running probabilistic model to predict binding sites of DNA binding proteins in NOMe-Seq data

Function for running probabilistic model to predict binding sites of DNA binding proteins in NOMe-Seq data

predict_footprints(
  data,
  footprint_models,
  bgprotectprob,
  bgcoverprior,
  ncpu = 1L,
  bound_fit_tol = 1e-05,
  bound_min_fderiv_val = 1e-05,
  bound_max_steps = 1000,
  run_priorEM = FALSE,
  priorEM_fit_tol = 1e-05,
  priorEM_max_steps = 1000,
  verbose = F
)

Arguments

data
A matrix containing NOMe-Seq data. Rows represent reads (or fragments), columns represent positions in amplicon. NOTE that data must contain only positions of selected GpCs where '0' represent unprotected "C", '1' represent protected "C" and all other positions must be filled with "NA".

footprint_models
A list containing footprint models for proteins. Each element must have 3 slots: PROTECT_PROB - a numeric vector with probability to find protected 'C' within the footprint; COVER_PRIOR - prior coverage probability (numeric) reflecting what fraction of reads you expect to be covered by a footprint; NAME - name (character) of a model, e.g. "Nucleosome".

bgprotectprob
probability to find a protected 'C' in open (uprotected) regions

bgcoverprior
prior probability for percentage of all fragments to be in a free (unprotected, or background) state

ncpu
number of threads to use

bound_fit_tol
fitting tolerance for estimating initial values of partition sums

bound_min_fderiv_val
This value used in Haley's numerical method for solving equations and represent a minimal value of denominator in first derivative of a function

bound_max_steps
Maximum number of iterations for algorithm which estimates boundary values for partition sums

run_priorEM
If TRUE the function runs Expectation Maximization algorithm for fitting prior probabilities of binding objects. Not recommended for regions with transcription factor binding as the EM overestimates shorter footprints.

priorEM_fit_tol
Fitting tolerance for running prior EM

priorEM_max_steps
Maximum number of iterations in prior EM

verbose
verbose mode for bug fixing

Value

A list which contains 3 data frames: START_PROB - data frame with calculated start probabilities for each NOME-Seq fragment (column seq), each position in the amplicon (column pos) and each binding model, e.g. Nucleosome, background etc. These probabilities reflect how likely it is to find a start in each fragment and at each position of a certain bidning model.

COVER_PROB - data frame with calculated coverage probabilities for each NOME-Seq fragment (column seq), each position in the amplicon (column pos) and each binding model, e.g. Nucleosome, background etc. These probabilities reflect how likely it is that a certain position in an amplicon and certain fragment is covered by a certain bidning model.

SUMMARY - data frame with summary information for each binding model and for the whole amplicon.

Column Statistics contains the following information:

Prior - empirical prior probabilities for each model.

Expected number of sites - expected number of sites for each model, i.e. sum of all start probabilities.

Coverage - average coverage of the amiplicon.

Sites<0.5 - expected number of sites with start probabilities lower than 0.5.

Sites>=0.5 - expected number of sites with start probabilities higher than 0.5.

Positions<0.5 - number of positions with start probabilities lower than 0.5.

Positionds>= - number of positions with start probabilities higher than 0.5.

Examples

set.seed(3346)
nc <- 50
nr <- 50
rmatr <- matrix(data = as.integer(rnorm(nc * nr) >= 0.5),
                ncol = nc,nrow=nr)

## create dummy footprints
bg.pr <- 0.5
ft.pr <- 1-bg.pr
ft.len <- 15

## creating a list of binding models for nomeR
ftp.models <- list(list("PROTECT_PROB" = rep(0.99,ft.len),
                        "COVER_PRIOR" = ft.pr,
                        "NAME" = "FOOTPRINT"))

nomeR.out <- predict_footprints(data=rmatr,
                                footprint_models = ftp.models,
                                bgprotectprob = 0.05,
                                bgcoverprior = bg.pr)

#> Warning: matr does not have row names. Fill the row names with indexes.