Create initial values for footprint inference

This utility function randomly draws initial values for inference from prior distributions.

init_ftp_abundance_for_inference(
  dir_alpha = c(200, rep(1, 199)),
  nchains = 4,
  ftp_background_model = c("ftp_bg_fixed", "bg_fixed", "informative_prior"),
  background_model_params = list(bg_protect_prob_fixed = 0.01, bg_protect_min = 0,
    bg_protect_max = 0.4, bg_protect_alpha = 675, bg_protect_beta = 57126),
  ftp_model_params = list(ftp_protect_prob_fixed = 0.99, ftp_protect_min = 0.6,
    ftp_protect_max = 1, ftp_protect_alpha = 1, ftp_protect_beta = 0.01),
  delta_from_max_min = 0.001
)

Arguments

dir_alpha
vector containing parameters for dirichlet distribution which is used to model footprint abundances. Default vector assumes that around 50% of data is free of any footprints and all other footprints up-to 200bp have uninformative prior.

nchains
number of chains

ftp_background_model
type of model used for inference.

"informative_prior"

both parameters bg_protect_prob and footprint_protect_prob as well as footprint abundances are subject for inference;

"bg_fixed"

bg_protect_prob is fixed value defined by background_model_params[["bg_protect_prob_fixed"]] and footprint_protect_prob as well as footprint abundances are subject for inference;

"ftp_bg_fixed"

bg_protect_prob and footprint_protect_prob are fixed values defined by corresponding values in background_model_params and ftp_model_params and inference runs only for footprint abundances.

background_model_params
list containing parameters for background model which must contain following elements:

`bg_protect_prob_fixed`

constant value for model parameter "bg_protect_prob" used in "bg_fixed" and "ftp_bg_fixed" models.

`bg_protect_min`

minimum allowed value for background emission probability for 1.

`bg_protect_max`

maximum allowed value for background emission probability for 1.

`bg_protect_alpha`

alpha parameter for prior beta distribution for model parameter "bg_protect_prob".

`bg_protect_beta`

beta parameter for prior beta distribution for model parameter "bg_protect_prob".

ftp_model_params
list containing parameters for footprint model which must contain following elements:

`ftp_protect_prob_fixed`

constant value for model parameter "footprint_protect_prob" used in "bg_fixed" and "ftp_bg_fixed" models.

`ftp_protect_min`

minimum allowed value for footprint emission probability for 1.

`ftp_protect_max`

maximum allowed value for footprint emission probability for 1.

`ftp_protect_alpha`

alpha parameter for prior beta distribution for model parameter "footprint_protect_prob".

`ftp_protect_beta`

beta parameter for prior beta distribution for model parameter "footprint_protect_prob".

delta_from_max_min
small number to avoid setting parameter values to ranges of the allowed interval which leads to error in stan.

Value

list of length `nchains` containing initial values for `ftp_abundances` and `footprint_protect_prob`.

Examples

## parameters for dirichlet distribution
footprint_prior_diralphas <- c(10,rep(1,14))

#' ## get initial values for sampling for vb
stan_initvals <- init_ftp_abundance_for_inference(dir_alpha = footprint_prior_diralphas)